osa4 4.c + 4.d

+++ osa4/blogilista/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+      .find({}).populate('user', { username: 1, name: 1})
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -13,6 +16,12 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request,response) => {
     const body = request.body
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = request.user
 
     if (!body.title) {
         return response.status(400).json({
@@ -34,16 +43,35 @@ blogsRouter.post('/', async (request,response) => {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user._id
     })
 
-    const blogsResult = await newblog.save()
-    response.status(200).json(blogsResult.toJSON())
+    const savedBlog = await newblog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save({ validateModifiedOnly: true })
+    response.status(200).json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const blogs = await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blogT = await Blog.findById(request.params.id)
+    const blogID = request.params.id
+    const user = request.user
+    if(blogT.user.toString() === decodedToken.id.toString()){
+        user.blogs = user.blogs.filter((value,index)=>{
+            return value!==blogID
+        })
+        await user.save({ validateModifiedOnly: true })
+        const blogs = await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }else{
+        return response.status(401).json({ error: 'cannot delete someone elses post' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
